Remove unused FF_HOOK_EVENTS
