+++ src/config/settings.py
@@ -231,5 +231,3 @@ if allowed_csrf_origins:
         allowed_csrf_origins.strip()
         for allowed_csrf_origins in allowed_csrf_origins.split(",")
     ]
-
-FF_HOOK_EVENTS = os.getenv("FF_HOOK_EVENTS", "false").lower() == "true"
