+++ src/safe_apps/signals.py
@@ -1,7 +1,6 @@
 import logging
 from typing import Any
 
-from django.conf import settings
 from django.core.cache import caches
 from django.db.models.signals import (
     m2m_changed,
@@ -12,7 +11,7 @@ from django.db.models.signals import (
 )
 from django.dispatch import receiver
 
-from clients.safe_client_gateway import HookEvent, flush, hook_event
+from clients.safe_client_gateway import HookEvent, hook_event
 
 from .models import Feature, Provider, SafeApp, Tag
 
@@ -23,31 +22,21 @@ logger = logging.getLogger(__name__)
 def on_safe_app_update(sender: SafeApp, instance: SafeApp, **kwargs: Any) -> None:
     logger.info("Clearing safe-apps cache")
     caches["safe-apps"].clear()
-    if settings.FF_HOOK_EVENTS:
-        chain_ids = set(instance.chain_ids)
-        if instance.app_id is not None:  # existing SafeApp being updated
-            previous = SafeApp.objects.filter(app_id=instance.app_id).first()
-            if previous is not None:
-                chain_ids.update(previous.chain_ids)
-        for chain_id in chain_ids:
-            hook_event(
-                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-            )
-    else:
-        flush()
+    chain_ids = set(instance.chain_ids)
+    if instance.app_id is not None:  # existing SafeApp being updated
+        previous = SafeApp.objects.filter(app_id=instance.app_id).first()
+        if previous is not None:
+            chain_ids.update(previous.chain_ids)
+    for chain_id in chain_ids:
+        hook_event(HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id))
 
 
 @receiver(post_delete, sender=SafeApp)
 def on_safe_app_delete(sender: SafeApp, instance: SafeApp, **kwargs: Any) -> None:
     logger.info("Clearing safe-apps cache")
     caches["safe-apps"].clear()
-    if settings.FF_HOOK_EVENTS:
-        for chain_id in instance.chain_ids:
-            hook_event(
-                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-            )
-    else:
-        flush()
+    for chain_id in instance.chain_ids:
+        hook_event(HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id))
 
 
 @receiver(post_save, sender=Provider)
@@ -55,14 +44,11 @@ def on_safe_app_delete(sender: SafeApp, instance: SafeApp, **kwargs: Any) -> Non
 def on_provider_update(sender: Provider, instance: Provider, **kwargs: Any) -> None:
     logger.info("Clearing safe-apps cache")
     caches["safe-apps"].clear()
-    if settings.FF_HOOK_EVENTS:
-        for safe_app in instance.safeapp_set.all():
-            for chain_id in safe_app.chain_ids:
-                hook_event(
-                    HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-                )
-    else:
-        flush()
+    for safe_app in instance.safeapp_set.all():
+        for chain_id in safe_app.chain_ids:
+            hook_event(
+                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
+            )
 
 
 # pre_delete is used because on pre_delete the model still has safe_apps
@@ -72,14 +58,11 @@ def on_provider_update(sender: Provider, instance: Provider, **kwargs: Any) -> N
 def on_tag_update(sender: Tag, instance: Tag, **kwargs: Any) -> None:
     logger.info("Clearing safe-apps cache")
     caches["safe-apps"].clear()
-    if settings.FF_HOOK_EVENTS:
-        for safe_app in instance.safe_apps.all():
-            for chain_id in safe_app.chain_ids:
-                hook_event(
-                    HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-                )
-    else:
-        flush()
+    for safe_app in instance.safe_apps.all():
+        for chain_id in safe_app.chain_ids:
+            hook_event(
+                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
+            )
 
 
 @receiver(m2m_changed, sender=Tag.safe_apps.through)
@@ -89,17 +72,14 @@ def on_tag_chains_update(
     logger.info("TagChains update. Triggering CGW webhook")
     caches["safe-apps"].clear()
     if action == "post_add" or action == "post_remove":
-        if settings.FF_HOOK_EVENTS:
-            chain_ids = set()
-            for safe_app in SafeApp.objects.filter(app_id__in=pk_set):
-                for chain_id in safe_app.chain_ids:
-                    chain_ids.add(chain_id)
-            for chain_id in chain_ids:
-                hook_event(
-                    HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-                )
-        else:
-            flush()
+        chain_ids = set()
+        for safe_app in SafeApp.objects.filter(app_id__in=pk_set):
+            for chain_id in safe_app.chain_ids:
+                chain_ids.add(chain_id)
+        for chain_id in chain_ids:
+            hook_event(
+                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
+            )
 
 
 # pre_delete is used because on pre_delete the model still has safe_apps
@@ -109,14 +89,11 @@ def on_tag_chains_update(
 def on_feature_update(sender: Feature, instance: Feature, **kwargs: Any) -> None:
     logger.info("Feature update. Triggering CGW webhook")
     caches["safe-apps"].clear()
-    if settings.FF_HOOK_EVENTS:
-        for safe_app in instance.safe_apps.all():
-            for chain_id in safe_app.chain_ids:
-                hook_event(
-                    HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-                )
-    else:
-        flush()
+    for safe_app in instance.safe_apps.all():
+        for chain_id in safe_app.chain_ids:
+            hook_event(
+                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
+            )
 
 
 @receiver(m2m_changed, sender=Feature.safe_apps.through)
@@ -126,14 +103,11 @@ def on_feature_safe_apps_update(
     logger.info("FeatureSafeApps update. Triggering CGW webhook")
     caches["safe-apps"].clear()
     if action == "post_add" or action == "post_remove":
-        if settings.FF_HOOK_EVENTS:
-            chain_ids = set()
-            for safe_app in SafeApp.objects.filter(app_id__in=pk_set):
-                for chain_id in safe_app.chain_ids:
-                    chain_ids.add(chain_id)
-            for chain_id in chain_ids:
-                hook_event(
-                    HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
-                )
-        else:
-            flush()
+        chain_ids = set()
+        for safe_app in SafeApp.objects.filter(app_id__in=pk_set):
+            for chain_id in safe_app.chain_ids:
+                chain_ids.add(chain_id)
+        for chain_id in chain_ids:
+            hook_event(
+                HookEvent(type=HookEvent.Type.SAFE_APPS_UPDATE, chain_id=chain_id)
+            )
